Updated to Revision 2 version
Updated to Revision 2 version.

Significant changes to the documentation.

Made ready for CRAN.

+++ src/stablearma.cpp
@@ -21,24 +21,9 @@ unsigned int stablearma(arma::mat U, arma::mat V, arma::ivec m, arma::ivec w) {
 //   i is partnered with m(i), so w(m(i))=i
      if( U(i,0) > U(i,m(i)) ){return(0);}
      for (unsigned int j = 0u; j < static_cast<unsigned int>(V.n_rows); ++j) {
-       if( (j != (m(i)-1)) && U(i,j+1) > U(i,m(i)) && V(j,i+1) > V(j,w(j)) ){return(0);}
+       if( (j != ((unsigned)(m(i))-1u)) && U(i,j+1) > U(i,m(i)) && V(j,i+1) > V(j,w(j)) ){return(0);}
      }
    }
  }
-//// man's perspective
-// for (unsigned int j = 0u; j < static_cast<unsigned int>(V.n_rows); ++j) {
-//   if(w(j)==0){
-////   j is single
-//     for (unsigned int i = 0u; i < static_cast<unsigned int>(U.n_rows); ++i) {
-//       if(V(j,i+1) > V(j,0) && U(i,j+1) > U(i,m(i))){return(0);}
-//     }
-//   }else{
-////   j is partnered with w(j), so m(w(j))=j
-//     if( V(j,0) > V(j,w(j)) ){return(0);}
-//     for (unsigned int i = 0u; i < static_cast<unsigned int>(U.n_rows); ++i) {
-//       if((i != (w(j)-1)) && V(j,i+1) > V(j,w(j)) && U(i,j+1) > U(i,m(i)) ){return(0);}
-//     }
-//   }
-// }
  return(1);
 }
